CH.01-Servers/custom-handlers
--ASSIGNMENT--
Let's add a readiness endpoint to the Chirpy server! Readiness endpoints
are commonly used by external systems to check if our server is ready to
receive traffic.
The endpoint should be accessible at the /healthz path using any HTTP
method.
The endpoint should simply return a 200 OK status code indicating that
it has started up successfully and is listening for traffic. The
endpoint should return a Content-Type: text/plain; charset=utf-8 header,
and the body will contain a message that simply says "OK" (the text
associated with the 200 status code).
Later this endpoint can be enhanced to return a 503 Service Unavailable
status code if the server is not ready.

Signed-off-by: Sam Gabel <sam@gabelcloud.com>