CH.02-Routing/stateful-handlers
--ASSIGNMENT--
The product managers at Chirpy want to know how many requests are being
made to serve our homepage - in essence, they want to know how many
people are viewing the site!
They have asked for a simple HTTP endpoint they can hit to get the
number of requests that have been processed. It will return the count as
plain text in the response body.
For now, they just want the number of requests that have been processed
since the last time the server was started, we don't need to worry about
saving the data between restarts.

Signed-off-by: Sam Gabel <sam@gabelcloud.com>