CH.02-Routing/routing
--ASSIGNMENT--
Add explicit HTTP methods to our current 2 custom endpoints to only
allow for GET methods.
- /healthz
- /metrics
In general, a pattern looks something like this: [METHOD ][HOST]/[PATH].
When a request is made to one of these endpoints with a method other
than GET, the server should return a 405 (Method Not Allowed) response
(this is handled automatically!).

Signed-off-by: Sam Gabel <sam@gabelcloud.com>