Ch.03-Architecture/monoliths-and-decoupling
--ASSIGNMENT--
For now, Chirpy is technically a monolith. That said we are keeping all
the API logic decoupled in the sense that it will be served from its own
namespace (path prefix). We serve the website from the app path, and
we'll be serving the API from the /api path.
Let's move our non-website endpoints to the /api namespace in our
routing.
To do this, prepend /api to the beginning of each of our API endpoints,
e.g. /api/healthz , /api/metrics , /api/reset

Signed-off-by: Sam Gabel <sam@gabelcloud.com>