CH.03-Architecture/admin-namespace
--ASSIGNMENT--
Let's swap out the /metrics endpoint, which just returns plain text, for
an /admin/metrics (not under the /api namespace) endpoint that returns
HTML intended to be rendered in the browser. It should be accessible via
GET requests only.
Make sure you use the Content-Type header to set the response type to
text/html so that the browser knows how to render it.
Try loading http://localhost:8080/admin/metrics in your browser, and in
another tab load http://localhost:8080/app a few times. Refreshing the
admin page should show the updated count.

Signed-off-by: Sam Gabel <sam@gabelcloud.com>