CH.04-JSON/the-profane
--ASSIGNMENT--
We need to update the /api/validate_chirp endpoint to replace all
"profane" words with 4 asterisks: ****.
Assuming the length validation passed, replace any of the following
words in the Chirp with the static 4-character string ****:
- kerfuffle
- sharbert
- fornax
Be sure to match against uppercase versions of the words as well, but
not punctuation. "Sharbert!" does not need to be replaced, we'll
consider it a different word due to the exclamation point. Finally,
instead of the valid boolean, your handler should return the cleaned
version of the text in a JSON response.

Signed-off-by: Sam Gabel <sam@gabelcloud.com>