CH.05-Storage/storage
--ASSIGNMENT--
Our API needs to support the standard CRUD operations for "chirps". A
"chirp" is just a short message that a user can post to the API, like a
tweet. For now, we'll just be adding the POST and GET endpoints to
create and read chirps respectively.
Keep your entire "database" in a single file called database.json at the
root of your project. Make sure to ignore it in Git. Your server should
automatically create the file if it doesn't exist upon startup.

Signed-off-by: Sam Gabel <sam@gabelcloud.com>