CH.05-Storage/get
--ASSIGNMENT--
Add a new endpoint to your server that allows users to get a single
chirp by ID. The response should be a 200 status code and the chirp
resource.
The {chirpID} part of the path should be replaced with the ID of the
chirp. For example, if the user wants to get chirp 1, they should make a
request to /api/chirps/1.
If the chirp does not exist, the server should return a 404 status code.

Signed-off-by: Sam Gabel <sam@gabelcloud.com>