CH.06-Authentication/authentication-with-passwords
--ASSIGNMENT--
Update the POST /api/users endpoint, and create a new POST /api/login
endpoint.
POST /api/users:
- Hash the password using the bcrypt.GenerateFromPassword function.
Bcrypt is a secure hash function that is intended to be used with
passwords.
- Be sure to store the hashed password in the database as you create the
user.
- The response from this endpoint should not include the password.
POST /api/login:
- This endpoint should allow a user to login. In a future exercise this
endpoint will be used to give the user a token that they can use to make
authenticated requests, but for now, let's just make sure the password
validation is working.
- You'll need a new way to look up a user by their email address (you
don't have access to an ID here)! At the same time, it would be a good
idea to update your database's "create user" logic to ensure that 2
users can't be created with the same email address.

Signed-off-by: Sam Gabel <sam@gabelcloud.com>