CH.06-Authentication/authentication-with-jwt
--ASSIGNMENT--
Update the POST /api/login endpoint, and create a new PUT /api/users
endpoint.
POST /api/login:
- This endpoint should accept a new optional expires_in_seconds field in
the request body. This field is used in our JWT creation.
- On authenticated login, we want to respond with a newly created JWT.
- In order to create a secret for the JWT creation, we should use the
godotenv package and a .env file.
PUT /api/login
- This endpoint should update a user's email and password.
- This is our first authenticated endpoint, which means it will require
a JWT to be present in the request headers.
- When handling the PUT request we want to make sure that the JWT hasn't
expired and our issuer matches.
- We want to respond to a valid and authorized PUT request with the user
id and updated email, also making sure to reflect the changes in the
database.

Signed-off-by: Sam Gabel <sam@gabelcloud.com>