CH.06-Authentication/refresh-tokens
--ASSIGNMENT--
We need to:
- Update the POST /api/login endpoint
- Add the POST /api/refresh endpoint
- Add the POST /api/revoke endpoint
POST /api/login:
- This endpoint should now return the refresh_token as well.
- It should also probably write the refresh_token to the database as
well.
- Refresh tokens expire after 60 days and the expiration time is stored
in the database.
POST /api/refresh:
- This new endpoint does not accept a request body but does require a
refresh token to be present in the headers, in the same Authorization:
Bearer <token> format.
- We should return a new JWT if the refresh token matches one in the
database, and it is unexpired.
- The newly created JWT should expire in 1 hour.
POST /api/revoke:
- This new endpoint does not accept a request body but does require a
refresh token to be present in the headers, in the same Authorization:
Bearer <token> format.
- Revoke the token in the database that matches the token that was
passed in the header of the request.

Signed-off-by: Sam Gabel <sam@gabelcloud.com>