CH.07-Authorization/authorization
--ASSIGNMENT--
Update the POST /api/chirps endpoint to only allow users to create
chirps if they are authenticated. When they do create a chirp, the chirp
should be associated with the user who created it. Add a new author_id
field to your chirps and use it to store the user's ID.
Make sure that the GET /api/chirps endpoints return the author_id field
as well.

Signed-off-by: Sam Gabel <sam@gabelcloud.com>