CH.07-Authorization/delete-chirp
--ASSIGNMENT--
Add a new DELETE /api/chirps/{chirpID} route to your server that deletes
a chirp from the database.
This is an authenticated endpoint, so be sure to check the token in the
header. Only allow the deletion of a chirp if the user is the author of
the chirp. If they aren't return a 403 status code.
If the chirp is deleted successfully, return a 204 status code..
-----------
NOTE:
- for simplicity we set the chirp id to len(dbStructure.Chirps) + 1.
Though the HTTP tests do not test for this, consider what happens if a
user does not delete the most recent chirp. We will get an ID conflict!

Signed-off-by: Sam Gabel <sam@gabelcloud.com>