CH.08-Webhooks/webhooks
--ASSIGNMENT--
We recently rolled out a new feature called "Chirpy Red". It's a
membership program, and members of "Chirpy Red" get vaporware features
that we'll announce in the future.
Until then, the only thing that Chirpy Red members get is a special
badge on their profile page.
Chirpy uses "Polka" as its payment provider, and they send us webhooks
whenever a user upgrades their account to Chirpy Red. We need to mark
users as Chirpy Red members when we receive these webhooks.
If the user is upgraded successfully, the endpoint should respond with a
204 status code and an empty response body. If the user can't be found,
the endpoint should respond with a 404 status code.

Signed-off-by: Sam Gabel <sam@gabelcloud.com>