CH.08-Webhooks/api-keys
--ASSIGNMENT--
Anyone can send a request to our webhook handler, and we'll process it.
That means that if Chirpy users figured out our API documentation, they
could simply upgrade their account without paying!
Luckily, Polka has a solution for this: API keys. Polka provided us with
an API key, and if a request to our webhook handler doesn't use that API
key, we should reject the request. This ensures that only Polka can tell
us to upgrade a user's account.
If any events come in without that API key, you should immediately
respond with a 401 status code.

Signed-off-by: Sam Gabel <sam@gabelcloud.com>