CH.09-Documentation/documentation
--ASSIGNMENT--
One type of endpoint that's nearly impossible to interact with without
documentation is a plural GET endpoint, that is, an endpoint that
returns a list of resources.
That's because they often have different sorting, filtering, and
pagination features.
Update the GET /api/chirps endpoint. It should accept an optional query
parameter called author_id. If the author_id query parameter is
provided, the endpoint should return only the chirps for that author. If
the author_id query parameter is not provided, the endpoint should
return all chirps as it did before.
For example:
GET http://localhost:8080/api/chirps?author_id=1

Signed-off-by: Sam Gabel <sam@gabelcloud.com>